Merge Develop to master

See merge request incore/incore!57

+++ server/fragility-service/src/main/java/edu/illinois/ncsa/incore/service/fragility/models/PeriodStandardFragilityCurve.java
@@ -1,16 +1,14 @@
-/*
- * Copyright (c) 2017 University of Illinois and others.  All rights reserved.
+/*******************************************************************************
+ * Copyright (c) 2019 University of Illinois and others.  All rights reserved.
  * This program and the accompanying materials are made available under the
- * terms of the BSD-3-Clause which accompanies this distribution,
- * and is available at https://opensource.org/licenses/BSD-3-Clause
+ * terms of the Mozilla Public License v2.0 which accompanies this distribution,
+ * and is available at https://www.mozilla.org/en-US/MPL/2.0/
  *
  * Contributors:
  * Omar Elabd, Nathan Tolbert
  */
 
 package edu.illinois.ncsa.incore.service.fragility.models;
-
-import java.math.BigDecimal;
 
 public class PeriodStandardFragilityCurve extends StandardFragilityCurve {
     public double periodParam2;
@@ -22,7 +20,7 @@ public class PeriodStandardFragilityCurve extends StandardFragilityCurve {
         super();
     }
 
-    public PeriodStandardFragilityCurve(BigDecimal median, BigDecimal beta, FragilityCurveType curveType, String label,
+    public PeriodStandardFragilityCurve(double median, double beta, FragilityCurveType curveType, String label,
                                         int periodEqnType, double periodParam0, double periodParam1, double periodParam2) {
         super(median, beta, curveType, label);
