+++ server/fragility-service/src/main/java/edu/illinois/ncsa/incore/service/fragility/models/StandardFragilityCurve.java
@@ -1,8 +1,8 @@
-/*
- * Copyright (c) 2017 University of Illinois and others.  All rights reserved.
+/*******************************************************************************
+ * Copyright (c) 2019 University of Illinois and others.  All rights reserved.
  * This program and the accompanying materials are made available under the
- * terms of the BSD-3-Clause which accompanies this distribution,
- * and is available at https://opensource.org/licenses/BSD-3-Clause
+ * terms of the Mozilla Public License v2.0 which accompanies this distribution,
+ * and is available at https://www.mozilla.org/en-US/MPL/2.0/
  *
  * Contributors:
  * Omar Elabd, Nathan Tolbert
@@ -10,18 +10,16 @@
 
 package edu.illinois.ncsa.incore.service.fragility.models;
 
-import java.math.BigDecimal;
-
 public class StandardFragilityCurve extends FragilityCurve {
-    public BigDecimal median;
-    public BigDecimal beta;
+    public double median;
+    public double beta;
     public FragilityCurveType curveType;
 
     public StandardFragilityCurve() {
         super();
     }
 
-    public StandardFragilityCurve(BigDecimal median, BigDecimal beta, FragilityCurveType curveType, String label) {
+    public StandardFragilityCurve(double median, double beta, FragilityCurveType curveType, String label) {
         super(label);
 
         this.median = median;
